refactor(main): 메인페이지 컴포넌트 분리 및 styling 코드 수정
