+++ src/features/main/ui/section/index.ts
@@ -1,1 +1,3 @@
-export * from './RecentClipCardSection';
+export * from './DescriptionSection';
+export * from './RecentClipSection';
+export * from './FriendListSection';
