+++ src/pages/index.ts
@@ -1,9 +1,7 @@
 export * from './add';
 export * from './clip';
 export * from './friend';
-export * from './login';
 export * from './main';
 export * from './search';
 export * from './setting';
-export * from './signup';
 export * from './summary';
